[OpenGL] Add ambient lighting
First component of the Phong lighting model.
Ambient lighting: even when it is dark there is usually still some light somewhere in the world (the moon, a distant light)
so objects are almost never completely dark.
To simulate this we use an ambient lighting constant that always gives the object some color.