[OpenGL] Add specular lighting
Last component of the Phong lighting model.
Specular reflection, also known as regular reflection, is the mirror-like reflection of waves, such as light, from a surface.
In this process, each incident ray is reflected at the same angle to the surface normal as the incident ray,
but on the opposing side of the surface normal in the plane formed by incident and reflected rays.
The result is that an image reflected by the surface is reproduced in mirror-like (specular) fashion.